+++ src/Authentication/ContainerInterface.php
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types=1);
+
+namespace Firehed\API\Authentication;
+
+use Psr\Container\ContainerInterface as Base;
+
+/**
+ * SHOULD use FQCN for get/has - etc
+ */
+interface ContainerInterface extends Base
+{
+}
