+++ src/Authorization/Exception.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace Firehed\API\Authorization;
+
+use RuntimeException;
+
+class Exception extends RuntimeException
+{
+}
