--- src/Exceptions/AuthenticationException.php
@@ -1,10 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Firehed\API\Exceptions;
-
-use RuntimeException;
-
-class AuthenticationException extends RuntimeException
-{
-}
