+++ src/Interfaces/AuthenticatedEndpointInterface.php
@@ -3,7 +3,9 @@ declare(strict_types=1);
 
 namespace Firehed\API\Interfaces;
 
+use Firehed\API\Authentication\ContainerInterface;
+
 interface AuthenticatedEndpointInterface extends EndpointInterface
 {
-    public function setAuthentication(AuthenticationContainerInterface $auth);
+    public function setAuthentication(ContainerInterface $auth);
 }
