--- src/Interfaces/AuthenticationContainerInterface.php
@@ -1,13 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Firehed\API\Interfaces;
-
-use Psr\Container\ContainerInterface;
-
-/**
- * SHOULD use FQCN for get/has - etc
- */
-interface AuthenticationContainerInterface extends ContainerInterface
-{
-}
