+++ src/extras.php
@@ -3,23 +3,24 @@ declare(strict_types=1);
 
 namespace Firehed\API\Interfaces;
 
-use Firehed\API\Exceptions\AuthenticationException;
+use Firehed\API\Authentication;
+use Firehed\API\Authorization;
 use Psr\Http\Message\ServerRequestInterface;
 // These need to be organized or possibly split into their own repo
 
 interface BearerTokenProcessorInterface
 {
-    public function process(string $token): AuthenticationContainerInterface;
+    public function process(string $token): Authentication\ContainerInterface;
 }
 
-class BearerTokenAuthentication implements AuthenticationProviderInterface
+class BearerTokenAuthentication implements Authentication\ProviderInterface
 {
     public function __construct(BearerTokenProcessorInterface $consumer)
     {
         $this->consumer = $consumer;
     }
 
-    public function authenticate(ServerRequestInterface $request): AuthenticationContainerInterface
+    public function authenticate(ServerRequestInterface $request): Authentication\ContainerInterface
     {
         // error checking...
         list($_, $token) = explode(' ', $request->getHeaderLine('Authorization'), 2);
@@ -28,9 +29,9 @@ class BearerTokenAuthentication implements AuthenticationProviderInterface
     }
 }
 
-abstract class BasicAuthAuthentiation implements AuthenticationProviderInterface
+abstract class BasicAuthAuthentiation implements Authentication\ProviderInterface
 {
-    public function authenticate(ServerRequestInterface $request): AuthenticationContainerInterface
+    public function authenticate(ServerRequestInterface $request): Authentication\ContainerInterface
     {
         $user = $_SERVER['PHP_AUTH_USER'];
         $pass = $_SERVER['PHP_AUTH_PW'];
@@ -38,21 +39,21 @@ abstract class BasicAuthAuthentiation implements AuthenticationProviderInterface
     }
 }
 
-class ChainAuthenticationProvider implements AuthenticationProviderInterface
+class ChainAuthenticationProvider implements Authentication\ProviderInterface
 {
     private $providers;
 
-    public function addProvider(AuthenticationProviderInterface $provider): ChainAuthProvider
+    public function addProvider(Authentication\ProviderInterface $provider): ChainAuthenticationProvider
     {
         $this->providers[] = $provider;
         return $this;
     }
-    public function authenticate(ServerRequestInterface $request): AuthenticationContainerInterface
+    public function authenticate(ServerRequestInterface $request): Authentication\ContainerInterface
     {
         foreach ($this->providers as $provider) {
             try {
                 return $provider->authenticate($request);
-            } catch (AuthenticationException $e) {
+            } catch (Authentication\Exception $e) {
             }
         }
         // Maybe? This could be some wacky exception(previous) chain rather
